Another refactor - Correct the `slnx` - Apply ef bundle migration & use-seeding | use-seeding-async - Configure dotnet's tools in local
