+++ src/Services/PeopleService/DNP.PeopleService.Tests.xUnitV3/DatabaseMigrationBackgroundService.cs
@@ -1,7 +1,8 @@
-﻿
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
-namespace DNP.PeopleService.BackgroundServices;
+namespace DNP.PeopleService.Tests.xUnitV3;
 
 public class DatabaseMigrationBackgroundService(ILogger<DatabaseMigrationBackgroundService> logger, IServiceScopeFactory scopeFactory) : IHostedService
 {
@@ -10,7 +11,7 @@ public class DatabaseMigrationBackgroundService(ILogger<DatabaseMigrationBackgro
 
     public async Task StartAsync(CancellationToken cancellationToken)
     {
-        using var scope = this._scopeFactory.CreateAsyncScope();
+        using var scope = _scopeFactory.CreateAsyncScope();
         var serviceProvider = scope.ServiceProvider;
         var dbContext = serviceProvider.GetRequiredService<DbContext>();
 
@@ -19,7 +20,7 @@ public class DatabaseMigrationBackgroundService(ILogger<DatabaseMigrationBackgro
 
         if (!pendingChanges.Any())
         {
-            this._logger.LogWarning("Nothing new. Database is up to date !!!!");
+            _logger.LogWarning("Nothing new. Database is up to date !!!!");
             return;
         }
