+++ src/Services/PeopleService/DNP.PeopleService/Program.cs
@@ -1,18 +1,11 @@
-
-using DNP.PeopleService.BackgroundServices;
-using DNP.PeopleService.Domain;
 using DNP.PeopleService.Persistence;
-using Microsoft.EntityFrameworkCore;
 using System.Text.Encodings.Web;
 using System.Text.Unicode;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Additional configuration is required to successfully run gRPC on macOS.
-// For instructions on how to configure Kestrel and gRPC clients on macOS, visit https://go.microsoft.com/fwlink/?linkid=2099682
+builder.AddPersistence();
 
-// Add services to the container.
-builder.Services.AddGrpc();
 builder.Services
     .AddEndpointsApiExplorer()
     .AddControllers()
@@ -21,45 +14,14 @@ builder.Services
         json.JsonSerializerOptions.Encoder = JavaScriptEncoder.Create(UnicodeRanges.All);
         json.JsonSerializerOptions.PropertyNameCaseInsensitive = true;
         json.JsonSerializerOptions.AllowTrailingCommas = true;
-        //json.JsonSerializerOptions.Converters.Add(new DateTimeJsonConverter());
-        //json.JsonSerializerOptions.Converters.Add(new NullableDateTimeJsonConverter());
     });
 
-var configuration = builder.Configuration;
-var connectionString = configuration.GetConnectionString("default");
-builder.Services.AddDbContextPool<DbContext, PeopleDbContext>(db =>
+builder.Services.AddGrpc();
+builder.WebHost.ConfigureKestrel(options =>
 {
-    db.UseSqlServer(connectionString, options =>
-    {
-        options.EnableRetryOnFailure();
-        options.MigrationsAssembly(typeof(PeopleDbContext).Assembly.GetName().Name);
-    });
-
-    db.UseSeeding((dbContext, _) =>
-    {
-        var peopleDbContext = (PeopleDbContext)dbContext;
-        var defaultPerson = peopleDbContext.People.FirstOrDefault(p => p.Id == Person.Default.Id);
-        if (defaultPerson == null)
-        {
-            peopleDbContext.Add(Person.Default);
-            peopleDbContext.SaveChanges();
-        }
-        
-    });
-
-    db.UseAsyncSeeding(async (dbContext, _, cancellationToken) =>
-    {
-        var peopleDbContext = (PeopleDbContext)dbContext;
-        var defaultPerson = await peopleDbContext.People.FirstOrDefaultAsync(p => p.Id == Person.Default.Id, cancellationToken);
-        if (defaultPerson == null)
-        {
-            peopleDbContext.Add(Person.Default);
-            await peopleDbContext.SaveChangesAsync(cancellationToken);
-        }
-    });
+    options.AddServerHeader = false;
+    options.Configure(builder.Configuration.GetSection("Kestrel"));
 });
-
-builder.Services.AddHostedService<DatabaseMigrationBackgroundService>();
 
 var app = builder.Build();
