+++ src/Services/PeopleService/DNP.PeopleService/appsettings.json
@@ -3,7 +3,9 @@
     "LogLevel": {
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning",
-      "Microsoft.Hosting.Lifetime": "Error",
+      "Microsoft.AspNetCore.Server.Kestrel.Http": "Debug",
+      "Microsoft.AspNetCore.Server.Kestrel.Http2": "Debug",
+      "Microsoft.Hosting.Lifetime": "Information",
       "Microsoft.AspNetCore.DataProtection": "Warning",
       "Microsoft.AspNetCore.Authentication": "Warning",
       "Microsoft.IdentityModel": "Warning",
@@ -14,10 +16,20 @@
   "AllowedHosts": "*",
   "Kestrel": {
     "EndpointDefaults": {
-      "Protocols": "Http2"
+      "Protocols": "Http1"
+    },
+    "Endpoints": {
+      "Http": {
+        "Protocols": "Http1",
+        "Url": "http://*:5049"
+      },
+      "Http2": {
+        "Url": "http://*:5048",
+        "Protocols": "Http2"
+      }
     }
   },
   "ConnectionStrings": {
-    "Default": "Server=localhost,1433;Database=ExampleDb;User Id=sa;Password=P@ssw0rd!;Encrypt=False;MultipleActiveResultSets=True"
+    "Default": "Server=localhost,1433;Database=People-Service;User Id=sa;Password=P@ssword;Encrypt=False;MultipleActiveResultSets=True"
   }
 }
